Refactor RequestPermissions #35 comments added to PermissionStatus icons changed RequestPermission, RequestCameraPermission, RequestMicrophonePermission and RequestLocationPermission refactored ShowRationalContent and PermissionDeniedContent added
