+++ permissions/build.gradle
@@ -45,6 +45,8 @@ dependencies {
     implementation "androidx.compose.ui:ui:$rootProject.ext.versions.compose"
     implementation "androidx.compose.material:material:$rootProject.ext.versions.compose"
     implementation "androidx.activity:activity-compose:$rootProject.ext.versions.activityCompose"
+
+    debugImplementation "androidx.compose.ui:ui-tooling:$rootProject.ext.versions.compose"
 }
 
 afterEvaluate {
