+++ permissions/src/main/java/com/patochallen/permissions/common/PermissionContentWrapper.kt
@@ -1,0 +1,48 @@
+package com.patochallen.permissions.common
+
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.PaddingValues
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.material.Button
+import androidx.compose.material.MaterialTheme
+import androidx.compose.material.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.dp
+
+@Composable
+internal fun PermissionContentWrapper(
+    buttonText: String,
+    onClick: () -> Unit,
+    content: @Composable (() -> Unit)
+) {
+    Column {
+        Box(
+            modifier = Modifier
+                .fillMaxWidth()
+                .weight(1f),
+            contentAlignment = Alignment.Center
+        ) {
+            content()
+        }
+        Button(
+            onClick = onClick,
+            modifier = Modifier
+                .fillMaxWidth()
+                .padding(horizontal = 32.dp, vertical = 16.dp),
+            shape = CircleShape,
+            contentPadding = PaddingValues(15.dp)
+        ) {
+            Text(
+                text = buttonText.uppercase(),
+                style = MaterialTheme.typography.button,
+                fontWeight = FontWeight.SemiBold
+            )
+        }
+    }
+}
