+++ sample/src/main/java/com/patochallen/sample/ui/theme/Theme.kt
@@ -1,3 +1,5 @@
+@file:Suppress("MagicNumber")
+
 package com.patochallen.sample.ui.theme
 
 import androidx.compose.foundation.isSystemInDarkTheme
@@ -7,6 +9,7 @@ import androidx.compose.material.Typography
 import androidx.compose.material.darkColors
 import androidx.compose.material.lightColors
 import androidx.compose.runtime.Composable
+import androidx.compose.ui.graphics.Color
 
 @Composable
 fun PermissionsSampleTheme(
@@ -20,7 +23,7 @@ fun PermissionsSampleTheme(
     }
 
     MaterialTheme(
-        colors = colors,
+        colors = colors.copy(primary = Color(0xFFE17A3F)),
         typography = Typography(),
         shapes = Shapes(),
         content = content
